Move method table definition to com_extension.c
