--- ext/com_dotnet/com_com.stub.php
@@ -1,7 +1,0 @@
-<?php
-
-class com
-{
-    /** @param string|array|null $server_name */
-    public function __construct(string $module_name, $server_name = UNKOWN, int $codepage = CP_ACP, string $typelib = "") {}
-}
