--- ext/com_dotnet/com_com_arginfo.h
@@ -1,8 +1,0 @@
-/* This is a generated file, edit the .stub.php file instead. */
-
-ZEND_BEGIN_ARG_INFO_EX(arginfo_class_com___construct, 0, 0, 1)
-	ZEND_ARG_TYPE_INFO(0, module_name, IS_STRING, 0)
-	ZEND_ARG_INFO(0, server_name)
-	ZEND_ARG_TYPE_INFO(0, codepage, IS_LONG, 0)
-	ZEND_ARG_TYPE_INFO(0, typelib, IS_STRING, 0)
-ZEND_END_ARG_INFO()
