+++ ext/com_dotnet/com_dotnet.c
@@ -26,7 +26,6 @@
 # include "php_com_dotnet.h"
 # include "php_com_dotnet_internal.h"
 # include "Zend/zend_exceptions.h"
-# include "com_dotnet_arginfo.h"
 # include <mscoree.h>
 
 /* Since there is no official public mscorlib.h header file, and since
@@ -316,11 +315,6 @@ PHP_METHOD(dotnet, __construct)
 	}
 }
 /* }}} */
-
-const zend_function_entry php_com_dotnet_funcs[] = {
-	PHP_ME(dotnet, __construct, arginfo_class_dotnet___construct, ZEND_ACC_PUBLIC)
-	PHP_FE_END
-};
 
 void php_com_dotnet_mshutdown(void)
 {
