--- ext/com_dotnet/com_dotnet.stub.php
@@ -1,6 +1,0 @@
-<?php
-
-class dotnet
-{
-    public function __construct(string $assembly_name, string $datatype_name, int $codepage = CP_ACP) {}
-}
