--- ext/com_dotnet/com_dotnet_arginfo.h
@@ -1,7 +1,0 @@
-/* This is a generated file, edit the .stub.php file instead. */
-
-ZEND_BEGIN_ARG_INFO_EX(arginfo_class_dotnet___construct, 0, 0, 2)
-	ZEND_ARG_TYPE_INFO(0, assembly_name, IS_STRING, 0)
-	ZEND_ARG_TYPE_INFO(0, datatype_name, IS_STRING, 0)
-	ZEND_ARG_TYPE_INFO(0, codepage, IS_LONG, 0)
-ZEND_END_ARG_INFO()
