+++ ext/com_dotnet/com_variant.c
@@ -23,7 +23,6 @@
 #include "ext/standard/info.h"
 #include "php_com_dotnet.h"
 #include "php_com_dotnet_internal.h"
-#include "com_variant_arginfo.h"
 
 /* create an automation SafeArray from a PHP array.
  * Only creates a single-dimensional array of variants.
@@ -1119,8 +1118,3 @@ PHP_FUNCTION(variant_cast)
 	VariantClear(&vres);
 }
 /* }}} */
-
-const zend_function_entry php_com_variant_funcs[] = {
-	PHP_ME(variant, __construct, arginfo_class_variant___construct, ZEND_ACC_PUBLIC)
-	PHP_FE_END
-};
