--- ext/com_dotnet/com_variant.stub.php
@@ -1,6 +1,0 @@
-<?php
-
-class variant
-{
-    public function __construct($value = null, int $type = VT_EMPTY, int $codepage = CP_ACP) {}
-}
