--- ext/com_dotnet/com_variant_arginfo.h
@@ -1,7 +1,0 @@
-/* This is a generated file, edit the .stub.php file instead. */
-
-ZEND_BEGIN_ARG_INFO_EX(arginfo_class_variant___construct, 0, 0, 0)
-	ZEND_ARG_INFO(0, value)
-	ZEND_ARG_TYPE_INFO(0, type, IS_LONG, 0)
-	ZEND_ARG_TYPE_INFO(0, codepage, IS_LONG, 0)
-ZEND_END_ARG_INFO()
