+++ ext/com_dotnet/php_com_dotnet_internal.h
@@ -87,13 +87,13 @@ PHP_COM_DOTNET_API OLECHAR *php_com_string_to_olestring(char *string,
 
 
 /* com_com.c */
-const zend_function_entry php_com_com_funcs[];
 PHP_FUNCTION(com_event_sink);
 PHP_FUNCTION(com_create_guid);
 PHP_FUNCTION(com_print_typeinfo);
 PHP_FUNCTION(com_message_pump);
 PHP_FUNCTION(com_load_typelib);
 PHP_FUNCTION(com_get_active_object);
+PHP_METHOD(com, __construct);
 
 HRESULT php_com_invoke_helper(php_com_dotnet_object *obj, DISPID id_member,
 		WORD flags, DISPPARAMS *disp_params, VARIANT *v, int silent, int allow_noarg);
@@ -115,7 +115,6 @@ PHP_COM_DOTNET_API IDispatch *php_com_wrapper_export(zval *val);
 int php_com_persist_minit(INIT_FUNC_ARGS);
 
 /* com_variant.c */
-const zend_function_entry php_com_variant_funcs[];
 PHP_FUNCTION(variant_set);
 PHP_FUNCTION(variant_add);
 PHP_FUNCTION(variant_cat);
@@ -142,6 +141,7 @@ PHP_FUNCTION(variant_date_from_timestamp);
 PHP_FUNCTION(variant_get_type);
 PHP_FUNCTION(variant_set_type);
 PHP_FUNCTION(variant_cast);
+PHP_METHOD(variant, __construct);
 
 PHP_COM_DOTNET_API void php_com_variant_from_zval_with_type(VARIANT *v, zval *z, VARTYPE type, int codepage);
 PHP_COM_DOTNET_API void php_com_variant_from_zval(VARIANT *v, zval *z, int codepage);
@@ -149,7 +149,7 @@ PHP_COM_DOTNET_API int php_com_zval_from_variant(zval *z, VARIANT *v, int codepa
 PHP_COM_DOTNET_API int php_com_copy_variant(VARIANT *dst, VARIANT *src);
 
 /* com_dotnet.c */
-const zend_function_entry php_com_dotnet_funcs[];
+PHP_METHOD(dotnet, __construct);
 void php_com_dotnet_rshutdown(void);
 void php_com_dotnet_mshutdown(void);
